Moved `HoodieSparkTypeUtils` under `spark.sql` to be able to access some package-private API
